requirements.txt
Django~=5.0.1
pillow